# Speed up count over nullable columns