--- a/tests/performance/countIf.xml
+++ b/tests/performance/countIf.xml
@@ -1,3 +1,21 @@
 <test>
     <query>SELECT countIf(number % 2) FROM numbers(100000000)</query>
+
+    <query>
+        SELECT countIf(key IS NOT NULL)
+        FROM
+        (
+            SELECT materialize(toNullable(1)) AS key
+            FROM numbers(100000000)
+        )
+    </query>
+
+    <query>
+        SELECT countIf(key IS NOT NULL)
+        FROM
+        (
+            SELECT materialize(CAST(NULL, 'Nullable(Int8)')) AS key
+            FROM numbers(100000000)
+        )
+    </query>
 </test>