Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speed up count over nullable columns


Detailed description / Documentation draft:

Implements AggregateFunctionCountNotNullUnary::addBatchSinglePlace so it doesn't fallback to the base class implementation which won't use countBytesInFilter / countBytesInFilterWithNull to check the flag and null arrays .


SQL:
```sql
SELECT count(n)
FROM
(
    SELECT toNullable(number) AS n
    FROM system.numbers
    LIMIT 10000000000
)
```

* Master: Elapsed: 2.385 sec. Processed 9.71 billion rows, 77.70 GB (4.07 billion rows/s., 32.57 GB/s.)
* PR: Elapsed: 1.825 sec. Processed 9.95 billion rows, 79.59 GB (5.45 billion rows/s., 43.61 GB/s.)

Closes https://github.com/ClickHouse/ClickHouse/issues/31627